add: error handler middleware

+++ src/services/task.service.js
@@ -6,6 +6,12 @@ async function create(task) {
 }
 
 async function getAll(userId) {
+    if (!userId) {
+        const error = new Error();
+        error.message = 'Se debe especificar el Id del usuario';
+        error.status = 400;
+        throw error;
+    }
     const tasks = await TaskModel.find({
         owner: userId
     }).exec();
@@ -13,16 +19,34 @@ async function getAll(userId) {
 }
 
 async function getById(id) {
+    if (!id) {
+        const error = new Error();
+        error.message = 'Se debe especificar el Id de la tarea';
+        error.status = 400;
+        throw error;
+    }
     const task = await TaskModel.findById(id).exec();
     return task;
 }
 
 async function update(id, updatedTask) {
+    if (!id) {
+        const error = new Error();
+        error.message = 'Se debe especificar el Id de la tarea';
+        error.status = 400;
+        throw error;
+    }
     const updated = await TaskModel.findByIdAndUpdate(id, updatedTask, { new: true }).exec();
     return updated;
 }
 
 async function destroy(id) {
+    if (!id) {
+        const error = new Error();
+        error.message = 'Se debe especificar el Id de la tarea';
+        error.status = 400;
+        throw error;
+    }
     const result = await TaskModel.findByIdAndDelete(id).exec();
     return result;
 }
